[REFACTOR] runWithLock반환 타입을 void로 수정
